pyproject.toml
[tool.poetry]
name = "stocks-patterns"
version = "0.1.0"
description = "Searching for patterns on stock price charts"
repository = "https://github.com/fasthydra/stocks_patterns"
authors = ["stocks-patterns_team <fasthydra@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]


[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.0"
numpy = "^1.24.2"
h5py = "^3.8.0"
plotly = "^5.14.1"
scikit-learn = "^1.2.2"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
jupyterlab = "^3.6.3"
matplotlib = "^3.7.1"
pre-commit = "^3.2.2"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
apimoex = "^1.3.0"
tqdm = "^4.65.0"
click = "^8.1.3"
pathlib = "^1.0.1"
mypy = "^1.2.0"
types-requests = "^2.29.0.0"
dvc = {extras = ["s3"], version = "^2.56.0"}
tslearn = "^0.5.3.2"
typing = "3.7.4.3"
mlflow = "^2.3.2"
optuna = "^3.2.0"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
python-multipart = "^0.0.6"

[tool.poetry.scripts]
load_data = 'src.data.load_data_from_moex:load_data_from_apimoex'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[mypy]
ignore_missing_imports = "True"

[tool.black]
line-length = 79
target-version = ['py39']
